Treat nil argument as blank argument in .search, return an empty array
All query methods are now consistent in treating nil as a blank query
and returning an empty result. In the case of .search, an empty result
is an empty array.

Assigning an empty string to the argument variable in order to treat it
exactly the same as an empty string.